--- morphine/__init__.py
@@ -1,2 +1,0 @@
-from morphine.domain.container import Container
-from morphine.domain.provider import Provider
