+++ morphine/providers/singleton.py
@@ -1,17 +1,22 @@
-from morphine import Provider
+from typing import Optional
+
+from morphine.provider import Provider
 
 
-class Factory[T](Provider):
-    """Provides a new instance with each call."""
+class Singleton[T](Provider):
+    """Provides the same instance on each call."""
+    _instance: Optional[T] = None
 
     def __call__(self) -> T:
+        if self._instance is not None:
+            return self._instance
         evaluated_kwargs = {}
         for key, value in self._constructor_kwargs.items():
             if isinstance(value, Provider):
                 evaluated_kwargs[key] = value()
                 continue
             evaluated_kwargs[key] = value
-        return self._instance_type(
+        self._instance = self._instance_type(
             *tuple(
                 arg() if isinstance(arg, Provider) else arg
                 for arg in self._constructor_args
@@ -21,3 +26,7 @@ class Factory[T](Provider):
                 for key, value in self._constructor_kwargs.items()
             },
         )
+        return self._instance
+
+    def reset(self) -> None:
+        self._instance = None
